// src/types/purchase.type.ts
import { Product } from './product.type'

export type PurchaseStatus = -1 | 1 | 2 | 3 | 4 | 5

export type PurchaseListStatus = PurchaseStatus | 0
export interface Status {
  status: number | string
}
export interface Purchase {
  _id: string
  buy_count: number
  price: number
  price_before_discount?: number
  status?: number | string
  user: string
  product_id: string
  product: Product
  createdAt: string
  updatedAt: string
}

export interface PurchaseList {
  message: string
  data: Purchase[]
}

export interface ExtendedPurchase extends Purchase {
  disabled: boolean
  checked: boolean
}